feat(List): create generic List component
The component and test file have been extracted from the
CategorySectionList component.

The reason of this choice is that the list component will be reused
by other components, too.

Note that items are not passed as children, but as props: indeed, a such
items are wrapped by li element by this component.