feat(StyledProductInfo): add StyledProductInfo module exporting PriceContainer styled component
StyledProductInfo will contain styled components for the elements
defining info of the products.

As of now, PriceContainer is exported, which defines a styled element
containing the price and possibly the discounted price, computed from
the price and discountedPrice props.

PriceContainer has been used in ProductItem to render the price /
discountedPrice. This simpliflies the code in the module and allow to
reuse it in a future page describing the item in more detail.