feat(HeadingLevelContext): add HeadingLevelContext module
This module creates the HeadingLevelContext providing the current
heading level and defines the HeadingLevelContextProvider component that
creates a new Provider wrapping component that increments by one the
heading level within it.

The Heading component has been modified to get the heading context from
the current HeadingLevelContext when hLevel props is set to "auto", and
that context is defined. This is to achieve automatic heading level
setting.

See: https://sergiodxa.com/tutorials/keep-heading-levels-consistent-with-react-context