feat(SavedProductsContext): add SavedProductsContext module
This module creates the SavedProductsContextProvider element which
creates a state for handling the saved products:

Sample state:
{
  productList: {
     0: { id: 0,  inCart: 4, inWishlist: true, /* other properties */ },
     1: { id: 1, inCart: 0, inWishlist: true, /* other properties */ },
     2: { id: 2,  inCart: 1, inWishlist: false, /* other properties */ },
     3: { id: 3, , inCart: 0, inWishlist: false, /* other properties */ },
  },
  wishlist: [0, 1], // list of ids
  cart: [0, 2], // list of ids
  cartItems: 5, // products in cart
};

Reducer function savedProductsReducer is used to handle this state.

Then, the helper functions  isInWishlist(id), isInCart(id), wishlist,
cart, and cartItems (number of items in the cart), as well as a dispatch
function to edit the cart, are created and provided to the children
elements through context.