feat(useLocalStorage): add useLocalStorage module
The useLocalStorageReducer custom hook is defined, which defines a state
using useReducer hook. The initial key, reducer, and initial state are
passed as props.

Then, the state is initialized from the value stored in Local Storage
(if present), otherwise using the initial state prop.

St each state update, the value stored in Local storage is updated.