feat(CartProductItem): add CartProductItem component that renders the info of a product in the cart
The component renders a Thumbnail, a Title, a price per item (possibly
showing the discounted price), an AddToCart button (which will be
possibly substituted with a diffrerent component allowing editing the
number of items in the cart), and a Wishlist button.

The component is similar to ProductItem component, as it has about
the same structure: just a few items a are different (as well as the
style that will be applied next).