test(CustomNumericInput): check that CustomNumericInput components renders and works correctly
The component will be a controlled input component with numeric data
within a range.

The component is currently composed just of two buttons to increment /
decrement the value according to some given callbacks.
The input element to show the current value and edit it is not rendered
yet.

This is exactly what is implemented in EditInCartButton component from
StyledProductInfo module, except it is more generic.

Hence the test is the same.

The reason to include this in its own module is to reuse it in other
part of the app, as well as simplify the code in StyledProductInfo and
its test (not done yet).