test(ProductFetch): check that the ProductFetch component is rendered correctly
The component will be used to fetch the data for a product given its id.
During the fetching, a loading spinner is shown. If the fetching is
successful, a Product element is rendered (not implemented, yet).
Otherwise, an error is thrown.

The test is based on the test of ProductFetchList component, as the
component will be quite similar (it renders a Product component instead
of a ProductList component). Some simplifications are made because there
is no "empty-list case", and no resetOnFetch is supported, among others.