test(ProductFetch): add ProductFetch component
The component is used to fetch the data for a product given its id.
During the fetching, a loading spinner is shown. If the fetching is
successful, a Product element is rendered (not implemented, yet).
Otherwise, an error is thrown.

The component is based on ProductFetchList component, as it is quite
similar.