test(StyledProductInfo): add basic test for AddMultipleToCart component in StyledProductInfo module
The component will have the possibility to set the number of items to
add to the cart.

Next, an "Add to cart" button will be added, to actually add the
selected number of items to the cart. The funcitonality will be tested,
too.

The test is quite similar to that of EditItemsInCart component from the
same module, which has the same structure.