feat(StyledProductInfo): add AddMultipleToCart component to StyledProductInfo module
So far the component is not functional, yet.

The component will have the possibility to set the number of items to
add to the cart, and an "Add to cart" button will be added, too, to
actually add the selected number of items to the cart.

The structure of the component is similar to that of EditItemsInCart
component from the same module.